+++ allennlp/nn/parallel/__init__.py
@@ -1,0 +1,1 @@
+from allennlp.nn.parallel.ddp_wrappers import DdpWrapper